requirements.txt
langgraph==1.1.2
langchain-groq==1.1.2
langchain-google-genai==4.2.1
langchain-community==0.4.1
tavily-python==0.7.23
playwright==1.58.0
beautifulsoup4==4.14.3
python-dotenv==1.2.2
aiohttp==3.13.3
pydantic==2.12.5
streamlit==1.55.0
